+++ service/controller/resource/s3bucket/update.go
@@ -3,11 +3,53 @@ package s3bucket
 import (
 	"context"
 
+	"github.com/aws/aws-sdk-go/aws"
+	"github.com/aws/aws-sdk-go/service/s3"
 	"github.com/giantswarm/microerror"
 	"github.com/giantswarm/operatorkit/v5/pkg/resource/crud"
+
+	"github.com/giantswarm/aws-operator/service/controller/controllercontext"
+	"github.com/giantswarm/aws-operator/service/controller/key"
 )
 
 func (r *Resource) ApplyUpdateChange(ctx context.Context, obj, updateChange interface{}) error {
+	// update bucket tags
+	cr, err := key.ToCluster(ctx, obj)
+	if err != nil {
+		return microerror.Mask(err)
+	}
+	cc, err := controllercontext.FromContext(ctx)
+	if err != nil {
+		return microerror.Mask(err)
+	}
+
+	currentBuckets, err := toBucketState(updateChange)
+	if err != nil {
+		return microerror.Mask(err)
+	}
+
+	for _, bucket := range currentBuckets {
+		if r.includeTags {
+			tagSet, err := r.getS3BucketTags(ctx, cr)
+			if err != nil {
+				return microerror.Mask(err)
+			}
+
+			i := &s3.PutBucketTaggingInput{
+				Bucket: aws.String(bucket.Name),
+				Tagging: &s3.Tagging{
+					TagSet: tagSet,
+				},
+			}
+
+			_, err = cc.Client.TenantCluster.AWS.S3.PutBucketTagging(i)
+			if err != nil {
+				return microerror.Mask(err)
+			}
+		}
+		r.logger.Debugf(ctx, "S3 bucket %#q tags updated", bucket.Name)
+	}
+
 	return nil
 }
 
@@ -29,7 +71,9 @@ func (r *Resource) NewUpdatePatch(ctx context.Context, obj, currentState, desire
 	return patch, nil
 }
 
-// newUpdateChange is a no-op because S3 buckets are not updated.
+// newUpdateChange returns all buckets in order to update tags on all buckets
 func (r *Resource) newUpdateChange(ctx context.Context, obj, currentState, desiredState interface{}) (interface{}, error) {
-	return nil, nil
+	currentBuckets, err := toBucketState(currentState)
+
+	return currentBuckets, err
 }