+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [10.9.1] - 2021-09-29
+
 ### Added
 
 - Add cloud tags propagation to S3 buckets.
@@ -482,7 +484,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 
-[Unreleased]: https://github.com/giantswarm/aws-operator/compare/v10.9.0...HEAD
+[Unreleased]: https://github.com/giantswarm/aws-operator/compare/v10.9.1...HEAD
+[10.9.1]: https://github.com/giantswarm/aws-operator/compare/v10.9.0...v10.9.1
 [10.9.0]: https://github.com/giantswarm/aws-operator/compare/v10.8.0...v10.9.0
 [10.8.0]: https://github.com/giantswarm/aws-operator/compare/v10.7.1...v10.8.0
 [10.7.1]: https://github.com/giantswarm/aws-operator/compare/v10.7.0...v10.7.1